streaming_decoder/requirements.txt
numpy==1.22.3
librosa==0.10.0.post2
scipy==1.10.1
soundfile==0.12.1
matplotlib
autobahn
twisted
h5py
pyyaml
omegaconf
pydantic
hydra-core==1.3.2
websockets==11.0.2
requests==2.25.1
tqdm
importlib_resources
pyopenssl
ipython
flask
flask-restful
pyvi
onnxruntime==1.14.1
regex
flashlight-text
rapidfuzz==2.13.7
fuzzysearch==0.7.3
webrtcvad==2.0.10
regex
uvloop
pydub==0.25.1
speechbrain==1.0.2